requirements/production.txt
Django==1.3.5
django-pagination==1.0.7
wsgiref==0.1.2
django-taggit==0.9.3
-e git+https://github.com/django-mptt/django-mptt.git@0af02a95877041b2fd6d458bd95413dc1666c321#egg=django-mptt
-e git+https://github.com/ahupp/python-magic.git@0d8f718f786a6ed7691fb3d9af726e395ac75ac3#egg=python-magic
slate==0.3
ghostscript==0.4.1
pdfminer==20110227
APScheduler==2.0.3
Pillow==1.7.8
cssmin==0.1.4
django-compressor==1.1.1
-e git+https://github.com/rosarior/django-sendfile.git#egg=django-sendfile
djangorestframework==0.2.3
South==0.7.6
#https://github.com/rosarior/python-gnupg/zipball/0.2.8
python-hkp==0.1.3
requests==0.14.1
pbs==0.105
psutil==0.5.1
GitPython==0.3.2.RC1
PyYAML==3.10
unicode-slugify==0.1
